requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
b2sdk==2.8.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
boto3==1.36.12
botocore==1.36.12
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
deprecation==2.1.0
Django==5.1.5
django-environ==0.12.0
django-storages==1.14.4
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.55.8
frozenlist==1.5.0
geopandas==1.0.1
gotrue==2.11.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.8
logfury==1.0.1
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
postgrest==0.19.3
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
realtime==2.2.0
requests==2.32.3
s3transfer==0.11.2
scipy==1.15.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==75.8.0
shapely==2.0.7
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
Werkzeug==3.1.3
whitenoise==6.8.2
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
